requirements.txt
fastmcp
pandas
openpyxl
xlrd
click
httpx
tuik-scraper
Flask
PyJWT
cryptography